fix: update user_id to BIGINT and handle foreign key constraint errors during support ticket creation

+++ internal/repository/postgres/migrations/000012_create_support_ticket.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE support_ticket (
     id          SERIAL PRIMARY KEY,
-    user_id     INT NOT NULL REFERENCES "user"(id),
+    user_id     BIGINT NOT NULL REFERENCES "user"(id),
     category    VARCHAR(50) NOT NULL,
     status      VARCHAR(50) NOT NULL DEFAULT 'open',
     title       VARCHAR(255) NOT NULL,
